+++ src/lib.rs
@@ -1,9 +1,8 @@
 // Copyright © 2025 Stephan Kunz
 // Copyright © of crate 'xml_writer' Piotr Zolnierek
 #![no_std]
-
 #![doc = include_str!("../README.md")]
 
 mod woxml;
 
-pub use woxml::XmlWriter;
+pub use woxml::{Error, XmlWriter};